Add files via upload
method which takes two parameters true or false and returns true or false upon the following condition: The parameter weekday is true if it is a weekday, and the parameter vacation is true if we are on vacation. We sleep in if it is not a weekday or we're on vacation. Return true if we sleep in.